Added zip function and testing (WIP)
The only active test currently checks that an EPUB file remains valid
after an unzip and zip operation. This test uses a Docker image
created from a fork (https://github.com/carlosfy/epubcheck) of the
EPUBCheck project (https://github.com/w3c/epubcheck).
Therefore, Docker must be installed, and the user must be logged in
for the test to run.

Currently, the tests do not pass. I believe this is because the EPUB
format requires a specific order when zipping, but so far, I'm zipping
without considering the order.

Additionally, it is difficult to compare two EPUB files because the
 zip operation can be noisy, making bit-by-bit comparison unreliable.