Closes #8: Refactor mock_server into function and add tests
Created a function that sets up a mock server so it can be used
both in tests and in a script.

The script will be useful later when we implement the ability to
perform changes in an EPUB. Instead of calling the DeepL API to test
the algorithms, we would use this mock server.